The refactor
