+++ src/main/java/frc/robot/Constants.java
@@ -5,7 +5,7 @@
 package frc.robot;
 
 import edu.wpi.first.math.util.Units;
-import frc.robot.subsystems.ShooterSubsystem.SpeakerConfig;
+import frc.robot.subsystems.shooterSubsystem.ShooterSubsystem.SpeakerConfig;
 
 /**
  * The Constants class provides a convenient place for teams to hold robot-wide numerical or boolean
@@ -229,13 +229,13 @@ public final class Constants {
   }
 
   public static class enabledSubsystems {
-    public static final boolean intakeEnabled = true;
-    public static final boolean drivetrainEnabled = true;
-    public static final boolean visionEnabled = true;
-    public static final boolean usingPhoton = true;
-    public static final boolean elvEnabled = true;
+    public static final boolean intakeEnabled = false;
+    public static final boolean drivetrainEnabled = false;
+    public static final boolean visionEnabled = false;
+    public static final boolean usingPhoton = false;
+    public static final boolean elvEnabled = false;
     public static final boolean signalEnabled = false;
-    public static final boolean shooterEnabled = true;
+    public static final boolean shooterEnabled = false;
     public static final boolean triggerEnabled = false;
     public static final boolean turretEnabled = false;
   }
