+++ src/main/java/frc/robot/subsystems/vision/LimelightSubsystem.java
@@ -10,6 +10,7 @@ import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.networktables.NetworkTableInstance;
 import edu.wpi.first.wpilibj.Timer;
+import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import frc.robot.Robot;
 import frc.robot.stateManagement.AllianceColor;
@@ -18,7 +19,7 @@ import frc.robot.utilities.DebugEntry;
 import frc.robot.utilities.LimelightHelpers;
 import java.util.function.BiConsumer;
 
-public class LimelightSubsystem extends VisionSubsystem {
+public class LimelightSubsystem extends SubsystemBase implements VisionSubsystem {
   private LimelightHelpers.LimelightResults results;
 
   private int measurementsUsed = 0;
